storage: Move intent resolution methods to their own file
